# Conflicts:
#	back-end/app/Http/Controllers/GroupController.php
#	back-end/app/Models/Group.php
#	back-end/composer.lock
#	back-end/package-lock.json
#	front-end/.gitignore
#	front-end/package-lock.json
#	front-end/package.json
#	front-end/src/App.jsx
#	front-end/src/components/ui/buttons.jsx

+++ back-end/app/Http/Controllers/GroupController.php
@@ -7,6 +7,7 @@ use App\Models\MainTask;
 use App\Models\User;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class GroupController extends Controller
 {
@@ -15,7 +16,13 @@ class GroupController extends Controller
      */
     public function index()
     {
-        return Group::all();
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+//        echo JWTAuth::parseToken()->authenticate()->id;
+//        echo $userId;
+
+        return Group::with('users')->whereHas('users', function ($query) use ($userId) {
+            $query->where('users.id', $userId);
+        })->get();
     }
 
 
@@ -26,6 +33,7 @@ class GroupController extends Controller
                 return response(['error' => 'you are stupid'], 404);
             }
 
+            $userId = JWTAuth::parseToken()->authenticate()->id;
             //save the uploaded file to the server
             $imagePath = null;
             if ($request->hasFile('image')) {
@@ -36,7 +44,6 @@ class GroupController extends Controller
                 'name' => $request->name,
                 'description' => $request->description,
                 'image' => $imagePath,
-                'user_id' => $request->user_id
             ]);
             $group->save();
 
@@ -53,10 +60,24 @@ class GroupController extends Controller
             }
 
             return $group;
+        } catch (ModelNotFoundException $e) {
+            return response(['error' => $e], 500);
+        }
+    }
 
-        } catch (\Exception $e) {
-            // This will now catch ALL database errors and send the exact message to React!
-            return response(['error' => $e->getMessage()], 500);
+
+    /**
+     * Display the specified resource.
+     */
+    public function show(string $id)
+    {
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+        $group = Group::query()->findOrFail($id);
+        // todo aanpassen naar admin ipv user_id
+        if ($group->user_id === $userId) {
+            return $group;
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
         }
     }
 
@@ -66,22 +87,26 @@ class GroupController extends Controller
     public function edit(string $id, request $request)
     {
         try {
-            echo $request->role;
-            if (!$request->name || !$request->role || !$request->user_id) {
+
+            if (!$request->name || !$request->user_id) {
                 return response(['error' => 'you are stupid'], 404);
             }
             $group = Group::query()->findOrFail($id);
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+            // todo aanpassen naar admin ipv user_id
+            if ($group->user_id = $userId) {
 
-            $group->update([
-                'name' => $request->name ?? $group->name,
-                'description' => $request->description ?? $group->description,
-            ]);
-            $group->users()->updateExistingPivot($request->user_id, [
-                'role' => $request->role ?? $group->role
-            ]);
-
-
-            return $group;
+                $group->update([
+                    'name' => $request->name ?? $group->name,
+                    'description' => $request->description ?? $group->description,
+                ]);
+                $group->users()->updateExistingPivot($userId, [
+                    'role' => $request->role ?? $group->role
+                ]);
+                return $group;
+            } else {
+                return response()->json(['error' => 'you are not authorized'], 403);
+            }
         } catch (ModelNotFoundException $e) {
             return response(['error' => $e], 500);
         }
@@ -93,8 +118,15 @@ class GroupController extends Controller
      */
     public function delete(string $id)
     {
+        $userId = JWTAuth::parseToken()->authenticate()->id;
         $group = Group::query()->findOrFail($id);
-        $group->delete();
-        return $group;
+        // todo aanpassen naar admin ipv user_id
+        if ($group->user_id === $userId) {
+
+            $group->delete();
+            return $group;
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 }
